requirements.txt
asgiref==3.4.1
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2
djangorestframework==3.12.4
django-cors-headers
mysqlclient>=2.0
